+++ CHANGELOG.md
@@ -5,12 +5,24 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [1.14.5]
+## [1.15.1]
 
 ### Added
 
 - Currency to notification
  
+## [1.15.0]
+
+### Changed
+
+- Redesign & WCAG (accessibility) improvements in checkout related views
+
+## [1.14.5]
+
+### Changed
+
+- Transaction ID from order
+
 ## [1.14.4]
 
 ### Added